Main Menu fully interactable
The main menu is fully interactable. All buttons work when both clicked with mouse and keys. Indicator works perfectly.

Menu functionality will need to be implemented later when level 1 is loadable and the save/load system is implemented.

Any implementation of options will need to be done by someone else, if possible.